Misc things all linked together
Functions to go in and out of the kits room, but also a way to do other things if there is a major event. Detecting if there is one was obviously needed. Only added Protéger le Roi and Invasion for now.
Also added a function I forgot about, that changes tags/scoreboards when selecting kits.